Added a fragment with a list of users

+++ app/src/main/java/ru/netology/nework/data/api/dto/UserDto.kt
@@ -1,8 +1,7 @@
 package ru.netology.nework.data.api.dto
 
-
 data class UserDto(
-    val id: Int,
+    val id: Long,        // меняем Int на Long
     val login: String,
     val name: String,
     val avatar: String?
