+++ app/src/main/java/ru/netology/nework/model/User.kt
@@ -1,7 +1,7 @@
 package ru.netology.nework.model
 
 data class User(
-    val id: Int,
+    val id: Long,           // меняем Int на Long
     val login: String,
     val name: String,
     val avatar: String?
